Commands: Ignore Command Results
Difficult to do something with results consistently. if commands want to
show something, they should use `show` or `log` in code. If functions that are
also commands, want to return something, they can, without it beeing
displayed. If a function, that wants to return something if called as
function, but display something if invoked, make two functions out of
it.